requirements.txt
openai==0.28.0
flask==2.3.3
pandas==2.1.0
chromadb==0.4.9
tiktoken==0.4.0
PyPDF2==3.0.1
torch==2.0.1
transformers==4.33.1
Pillow==10.0.1
pdf2image==1.16.3
pytesseract==0.3.10
langchain==0.0.292